+++ op-node/flags/flags.go
@@ -2,6 +2,7 @@ package flags
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/urfave/cli"
 )
@@ -81,6 +82,13 @@ var (
 		Required: false,
 		Value:    4,
 	}
+	L1EpochPollIntervalFlag = cli.DurationFlag{
+		Name:     "l1.epoch-poll-interval",
+		Usage:    "Poll interval for retrieving new L1 epoch updates such as safe and finalized block changes. Disabled if 0 or negative.",
+		EnvVar:   prefixEnvVar("L1_EPOCH_POLL_INTERVAL"),
+		Required: false,
+		Value:    time.Second * 12 * 32,
+	}
 	LogLevelFlag = cli.StringFlag{
 		Name:   "log.level",
 		Usage:  "The lowest log level that will be output",
@@ -132,11 +140,26 @@ var (
 		Value:  6060,
 		EnvVar: prefixEnvVar("PPROF_PORT"),
 	}
-
 	SnapshotLog = cli.StringFlag{
 		Name:   "snapshotlog.file",
 		Usage:  "Path to the snapshot log file",
 		EnvVar: prefixEnvVar("SNAPSHOT_LOG"),
+	}
+	HeartbeatEnabledFlag = cli.BoolFlag{
+		Name:   "heartbeat.enabled",
+		Usage:  "Enables or disables heartbeating",
+		EnvVar: prefixEnvVar("HEARTBEAT_ENABLED"),
+	}
+	HeartbeatMonikerFlag = cli.StringFlag{
+		Name:   "heartbeat.moniker",
+		Usage:  "Sets a moniker for this node",
+		EnvVar: prefixEnvVar("HEARTBEAT_MONIKER"),
+	}
+	HeartbeatURLFlag = cli.StringFlag{
+		Name:   "heartbeat.url",
+		Usage:  "Sets the URL to heartbeat to",
+		EnvVar: prefixEnvVar("HEARTBEAT_URL"),
+		Value:  "https://heartbeat.bedrock-goerli.optimism.io",
 	}
 )
 
@@ -154,6 +177,7 @@ var optionalFlags = append([]cli.Flag{
 	VerifierL1Confs,
 	SequencerEnabledFlag,
 	SequencerL1Confs,
+	L1EpochPollIntervalFlag,
 	LogLevelFlag,
 	LogFormatFlag,
 	LogColorFlag,
@@ -165,6 +189,9 @@ var optionalFlags = append([]cli.Flag{
 	PprofAddrFlag,
 	PprofPortFlag,
 	SnapshotLog,
+	HeartbeatEnabledFlag,
+	HeartbeatMonikerFlag,
+	HeartbeatURLFlag,
 }, p2pFlags...)
 
 // Flags contains the list of configuration options available to the binary.