+++ packages/contracts-bedrock/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @eth-optimism/contracts-bedrock
 
+## 0.6.2
+
+### Patch Changes
+
+- 651a2883: Make spacer variables private in the Bridge and Messenger contracts so that they cannot be accessed in child contracts.
+
+## 0.6.1
+
+### Patch Changes
+
+- 85232179: Add CrossDomainOwnable contracts
+- 593f1cfb: Removes the blockedSystemMessages mapping in favor of a simpler approach to preventing messages from being sent to internal system addresses.
+- f78eb056: Prevents v0 (legacy) messages from being relayed in the bedrock XDM.
+
+## 0.6.0
+
+### Minor Changes
+
+- 7fdc490c: Removes initializer from StandardBridge in favor of immutables
+- 3d228a0e: Updates the storage layout for the CrossDomainMessenger base contract to reduce diff with the existing system.
+
+### Patch Changes
+
+- 63ef1949: Delete hardhat genesis tasks
+- Updated dependencies [dbfea116]
+  - @eth-optimism/core-utils@0.10.0
+
 ## 0.5.4
 
 ### Patch Changes